backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
pymysql==1.1.0
sqlalchemy==2.0.23
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.1
python-dotenv==1.0.0 